feat: restructure whole codebase
