+++ .env.sample
@@ -1,0 +1,13 @@
+PORT=
+EMAIL_USER=
+
+# Client credentials for OAuth  
+CLIENT_ID=
+CLIENT_SECRET=
+REDIRECT_URI=
+REFRESH_TOKEN=
+ACCESS_TOKEN=
+
+# Credentials for WhatShapp API
+WHATSAPP_PHONE_ID=
+WHATSAPP_ACCESS_TOKEN=
