+++ utils/nodemailerConfig.js
@@ -1,0 +1,18 @@
+const nodemailer = require("nodemailer");
+
+// Function to Create Email Transporter
+async function createTransporter() {
+    return nodemailer.createTransport({
+      service: "gmail",
+      auth: {
+        type: "OAuth2",
+        user: process.env.EMAIL_USER,
+        clientId: process.env.CLIENT_ID,
+        clientSecret: process.env.CLIENT_SECRET,
+        refreshToken: process.env.REFRESH_TOKEN,
+        accessToken: process.env.ACCESS_TOKEN,
+      },
+    });
+  }
+
+module.exports = createTransporter;
